Namespace refactoring
