+++ CBinding/Parser/Function.cs
@@ -6,7 +6,7 @@ using GLib;
 using System.Runtime.InteropServices;
 using ICSharpCode.NRefactory.MonoCSharp;
 
-namespace CBinding
+namespace CBinding.Parser
 {
 	public class Function : Symbol
 	{
